Adjust min_ and max_size arguments
Now the min and max size refer to the number of coding amino acids in
the output coding sequence. That is, the STOP codon is not counted. The
START codon is counted, though.

How should the input `min_size=0` be handled when
`include_start_codon=True`? These are contradictory. If there is a START
codon, then the minimum number of codons is 1. Currently, I silently
ignore the 0, treating it as if it were 1 by still including the START
codon.